+++ core/test/Service/KRPC/KRPCTest.cs
@@ -30,7 +30,7 @@ namespace KRPC.Test.Service.KRPC
             Assert.AreEqual (5, services.ServicesList.Count);
 
             var service = services.ServicesList.First (x => x.Name == "KRPC");
-            Assert.AreEqual (69, service.Procedures.Count);
+            Assert.AreEqual (72, service.Procedures.Count);
             Assert.AreEqual (2, service.Classes.Count);
             Assert.AreEqual (1, service.Enumerations.Count);
 
@@ -74,6 +74,18 @@ namespace KRPC.Test.Service.KRPC
                     MessageAssert.HasParameters (proc, 1);
                     MessageAssert.HasParameter (proc, 0, typeof(ulong), "id");
                     MessageAssert.HasDocumentation (proc);
+                } else if (proc.Name == "HoldTick") {
+                    MessageAssert.HasNoReturnType (proc);
+                    MessageAssert.HasNoParameters (proc);
+                    MessageAssert.HasDocumentation (proc);
+                } else if (proc.Name == "ReleaseTick") {
+                    MessageAssert.HasNoReturnType (proc);
+                    MessageAssert.HasNoParameters (proc);
+                    MessageAssert.HasDocumentation (proc);
+                } else if (proc.Name == "NextTick") {
+                    MessageAssert.HasNoReturnType (proc);
+                    MessageAssert.HasNoParameters (proc);
+                    MessageAssert.HasDocumentation (proc);
                 } else if (proc.Name == "get_Clients") {
                     MessageAssert.HasReturnType (proc, typeof(IList<Tuple<byte[],string,string>>));
                     MessageAssert.HasNoParameters (proc);
@@ -97,7 +109,7 @@ namespace KRPC.Test.Service.KRPC
                 }
                 foundProcedures++;
             }
-            Assert.AreEqual (12, foundProcedures);
+            Assert.AreEqual (15, foundProcedures);
 
             bool foundEnumeration = false;
             foreach (var enumeration in service.Enumerations) {